test: shuffle files a bit
This should hopefully reduce clutter in the test/ directory.

Side note: if I leave the __init__.py file in the new py/ directory,
pytest fails with import errors. To make it work, I need to either
delete it or keep the __init__.py file in both test/ and py/. No idea
why.