use __WNOTHREAD w/ waitpid
I learned about this flag on my previous job; basically, it should be
enabled by default IMO, so that the thread doesn't receive signals from
other threads' children.

Here, it doesn't matter too much, since we're waiting for a specific
child. However, if we were to use waitpid(-1, ...), it would be
essential to use this flag. Still, even here, it's good to have this on.